Spell and grammar check
Cleared up some spelling and grammar errors, and made some sections more readable.
Important: Matthew, please look at lines 55, 63, 72, and 105, I changed the if conditional to have the words "is less than" instead of just <, feel free to change it back, it is merely a suggestion.
At lines 77, it talks about installing the shield. I believe that part was already covered in the previous tutorial.